educational programs search, edit, add, delte funtions has been activated!

+++ Larvel/app/Http/Controllers/programController.php
@@ -92,41 +92,41 @@ class programController extends Controller
         }
         elseif($request->path() === 'admin/pdcProgramList'){
             // return "akdsfjaksjflj";
-            $validate = $request->validate([
-                'name' => 'bail|required|string|max:100',
-                'type' => 'bail|required|string|in:ورکشاپ,سیمینار,سمفوزیم,کنفرانس',
-                'sponsor' => 'bail|required|string|max:30',
-                'supporter' => 'bail|required|string|max:30',
-                'manager' => 'bail|required|string|max:30',
-                'facilitator' => 'bail|nullable|string|max:30',
-                'info_mobile_number' => 'bail|required|string|max:13',
-                'participant_amount' => 'bail|required|integer|between:1,1000',
-                'fund' => 'bail|required|integer',
-                'fund_type' => 'bail|required|string|in:افغانۍ,ډالر',
-                'fee_able' => 'bail|required|integer|in:0,1',
-                'fee' => 'bail|integer|required_unless:fee_able,=,1|integer|gte:1',
-                'fee_unit' => 'bail|string|required_if:fee_able,=,1|in:افغانۍ,ډالر',
-                'campus_name' => 'bail|required|string|max:30',
-                'block_name' => 'bail|required|string|max:30',
-                'block_number' => 'bail|required|integer|between:1,30',
-                'room_number' => 'bail|required|integer|between:1,30',
-                'start_date' => 'bail|required|date',
-                'end_date' => 'bail|required|date',
-                // 'start_day' => 'bail|required|integer|between:1,31',
-                // 'end_day' => 'bail|required|integer|between:1,31',
-                // 'start_time' => 'bail|required|date_format:H:i',
-                // 'end_time' => 'bail|required|date_format:H:i',
-                // 'days_duration' => 'bail|required|integer',
-                'program_description' => 'bail|required|string|max:2000',
-            ]);
-            // return $request->fee_able;
-            if($request->fee_able == 1){
-                // return "lsdflds";
-                $validate = $request->validate([
-                    'fee' => 'bail|integer|required',
-                    'fee_unit' => 'bail|string|required|in:افغانۍ,ډالر',
-                ]);
-            }
+            // $validate = $request->validate([
+            //     'name' => 'bail|required|string|max:100',
+            //     'type' => 'bail|required|string|in:ورکشاپ,سیمینار,سمفوزیم,کنفرانس',
+            //     'sponsor' => 'bail|required|string|max:30',
+            //     'supporter' => 'bail|required|string|max:30',
+            //     'manager' => 'bail|required|string|max:30',
+            //     'facilitator' => 'bail|nullable|string|max:30',
+            //     'info_mobile_number' => 'bail|required|string|max:13',
+            //     'participant_amount' => 'bail|required|integer|between:1,1000',
+            //     'fund' => 'bail|required|integer',
+            //     'fund_type' => 'bail|required|string|in:افغانۍ,ډالر',
+            //     'fee_able' => 'bail|required|integer|in:0,1',
+            //     'fee' => 'bail|integer|required_unless:fee_able,=,1|integer|gte:1',
+            //     'fee_unit' => 'bail|string|required_if:fee_able,=,1|in:افغانۍ,ډالر',
+            //     'campus_name' => 'bail|required|string|max:30',
+            //     'block_name' => 'bail|required|string|max:30',
+            //     'block_number' => 'bail|required|integer|between:1,30',
+            //     'room_number' => 'bail|required|integer|between:1,30',
+            //     'start_date' => 'bail|required|date',
+            //     'end_date' => 'bail|required|date',
+            //     // 'start_day' => 'bail|required|integer|between:1,31',
+            //     // 'end_day' => 'bail|required|integer|between:1,31',
+            //     // 'start_time' => 'bail|required|date_format:H:i',
+            //     // 'end_time' => 'bail|required|date_format:H:i',
+            //     // 'days_duration' => 'bail|required|integer',
+            //     'program_description' => 'bail|required|string|max:2000',
+            // ]);
+            // return $request->start_date;
+            // if($request->fee_able == 1){
+            //     // return "lsdflds";
+            //     $validate = $request->validate([
+            //         'fee' => 'bail|integer|required',
+            //         'fee_unit' => 'bail|string|required|in:افغانۍ,ډالر',
+            //     ]);
+            // }
             // return $request->start_date;
             $start = Carbon::parse($request->input('start_date'));
             $end = Carbon::parse($request->input('end_date'));
@@ -232,14 +232,55 @@ class programController extends Controller
         if($request->path() === 'admin/pdcProgramList/'.$id)
         {
             // return "i am edit.";
-            
+            // return $request->fee_able;
+            $validate = $request->validate([
+                'name' => 'bail|required|string|max:100',
+                'type' => 'bail|required|string|in:ورکشاپ,سیمینار,سمفوزیم,کنفرانس',
+                'sponsor' => 'bail|required|string|max:30',
+                'supporter' => 'bail|required|string|max:30',
+                'manager' => 'bail|required|string|max:30',
+                'facilitator' => 'bail|nullable|string|max:30',
+                'info_mobile_number' => 'bail|required|string|max:13',
+                'participant_amount' => 'bail|required|integer|between:1,1000',
+                'fund' => 'bail|required|integer',
+                'fund_type' => 'bail|required|string|in:افغانۍ,ډالر',
+                'fee_able' => 'bail|required|integer|in:0,1',
+                // 'fee' => 'bail|integer|required_unless:fee_able,=,1|integer|gte:1',
+                // 'fee_unit' => 'bail|string|required_if:fee_able,=,1|in:افغانۍ,ډالر',
+                'campus_name' => 'bail|required|string|max:30',
+                'block_name' => 'bail|required|string|max:30',
+                'block_number' => 'bail|required|integer|between:1,30',
+                'room_number' => 'bail|required|integer|between:1,30',
+                'start_date' => 'bail|required|date',
+                'end_date' => 'bail|required|date',
+                // 'start_day' => 'bail|required|integer|between:1,31',
+                // 'end_day' => 'bail|required|integer|between:1,31',
+                // 'start_time' => 'bail|required|date_format:H:i',
+                // 'end_time' => 'bail|required|date_format:H:i',
+                // 'days_duration' => 'bail|required|integer',
+                'program_description' => 'bail|required|string|max:2000',
+            ]);
+            if($request->fee_able == 1){
+                    // return ;
+                $validate = $request->validate([
+                   'fee' => 'bail|integer|required',
+                   'fee_type' => 'bail|string|required|in:افغانۍ,ډالر',
+                   'agenda.*' => 'bail|string|required',
+                   'facility.*' => 'bail|string|required',
+                   'result.*' => 'bail|string|required',
+                   'evaluation.*' => 'bail|string|required',
+                ]);
+            }
+            // if($request->result == null)
+            // {
+            //     return $request->result."sdlkfjsd";
+            // }
             $addProgram = Program::with('getResults', 'getFacilities', 'getAgendas', 'getEvaluations')->find($id);
             $addProgram->name = $request->name;
             $addProgram->type = $request->type;
             $addProgram->sponsor = $request->sponsor;
             $addProgram->supporter = $request->supporter;
             $addProgram->manager = $request->manager;
-            $addProgram->facilitator = $request->facilitator;
             $addProgram->participant_amount = $request->participant_amount;
             $addProgram->fund = $request->fund;
             $addProgram->info_mobile_number = $request->info_mobile_number;
@@ -255,41 +296,54 @@ class programController extends Controller
             $addProgram->block_number = $request->block_number;
             $addProgram->room_number = $request->room_number;
             $addProgram->save();
-            // this part belongs to Facility Model
-            Facility::where('program_id', $id)->delete();
-            foreach ($request->facility as  $value) {
-                $programFacility = new Facility;
-                $programFacility->facility = $value;
-                $programFacility->program_id = $id;
-                $programFacility->save();
-                // sleep(1);
+            if($request->facility != null)
+            {
+                // this part belongs to Facility Model
+                Facility::where('program_id', $id)->delete();
+                foreach ($request->facility as  $value) {
+                    $programFacility = new Facility;
+                    $programFacility->facility = $value;
+                    $programFacility->program_id = $id;
+                    $programFacility->save();
+                    // sleep(1);
+                }
             }
-            // this part belongs to Agenda Model
-            Agenda::where('program_id', $id)->delete();
-            foreach ($request->agenda as  $value) {
-                $programAgenda = new Agenda;
-                $programAgenda->agenda = $value;
-                $programAgenda->program_id = $id;
-                $programAgenda->save();
-                // sleep(1);
+            if($request->agenda != null)
+            {
+                // this part belongs to Agenda Model
+                Agenda::where('program_id', $id)->delete();
+                foreach ($request->agenda as  $value) {
+                    $programAgenda = new Agenda;
+                    $programAgenda->agenda = $value;
+                    $programAgenda->program_id = $id;
+                    $programAgenda->save();
+                    // sleep(1);
+                }
             }
-            // this part belongs to result Model
-            Result::where('program_id', $id)->delete();
-            foreach ($request->result as  $value) {
-                $programResult = new Result;
-                $programResult->result = $value;
-                $programResult->program_id = $id;
-                $programResult->save();
-                // sleep(1);
+            if($request->result != null)
+            {
+                // this part belongs to result Model
+                Result::where('program_id', $id)->delete();
+                foreach ($request->result as  $value) {
+                    $programResult = new Result;
+                    $programResult->result = $value;
+                    $programResult->program_id = $id;
+                    $programResult->save();
+                    // sleep(1);
+                }
             }
-            // this part belongs to evaluation Model
-            Evaluation::where('program_id', $id)->delete();
-            foreach ($request->evaluation as  $value) {
-                $programEvaluation = new Evaluation;
-                $programEvaluation->evaluation = $value;
-                $programEvaluation->program_id = $id;
-                $programEvaluation->save();
-                // sleep(1);
+
+            if($request->evaluation != null)
+            {
+                // this part belongs to evaluation Model
+                Evaluation::where('program_id', $id)->delete();
+                foreach ($request->evaluation as  $value) {
+                    $programEvaluation = new Evaluation;
+                    $programEvaluation->evaluation = $value;
+                    $programEvaluation->program_id = $id;
+                    $programEvaluation->save();
+                    // sleep(1);
+                }
             }
             
             // return "done";
